Add files via upload
Nowadays, the use of databases is adopted by companies in different industries such as technology, construction, manufacturing, healthcare, and retail. The needs of the database are growing rapidly especially in the retail industry because consumer purchase services and goods are changing every year. A great way to manage customer and product information is through the implementation of databases. Furthermore, databases can improve customer satisfaction since it can provide faster and convenient service in the companies’ daily operation; it can also increase business revenue by having access to various data types, to provide both overall and detailed analysis for future business strategy. 
 
Although the retail companies have different company scale and product types, the expectation and requirement for a suitable database are roughly the same. The logic of database designing for retail companies shall match the common business logic and needs in this industry. For customer management, the database will store customer information tidy and up to date, which can maintain a strong customer relationship to provide better service and promote future sales easily. For product management, the database will log and monitor all transactions and show product storage. For supplier and vendor management, the database will provide an effective way to record each product resource and previous or on-going activities. For internal uses, a powerful database also provides various access authority to users from different levels for better internal operation and management. Having a solid and reliable database to manage and allocate resources would benefit greatly to the company not only in increasing revenue, but also in establishing and maintaining a healthy relationship with customers, partners, and staff.
 
This project is to design a database foundation that can be used for the retail companies from small to large size, which will heavily involve customer, transaction, product, and related information in the database frame. This project includes system requirements, database design, database implementation, and feature and function introduction. 